Move struct declaration to header and update allowed tensor list
